change  model

+++ app/services/GraySystemService.scala
@@ -32,7 +32,7 @@ class GraySystemService @Inject() (graySystems: GraySystems)  {
   def listAllGraySystems: Future[Seq[GraySystem]] = {
     graySystems.listAll
   }
-  def listAllGraySystemsByConf(id: Long): Future[Seq[GraySystem]] = {
+  def listAllGraySystemsByConf(id: Int): Future[Seq[GraySystem]] = {
     graySystems.list(id)
   }
